Issue/24

+++ repository/tweet_repository.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"atwell/domain"
+	"atwell/infrastructure"
 	"time"
 
 	"gorm.io/gorm"
@@ -18,22 +19,40 @@ func NewMysqlTweetRepository(db *gorm.DB) domain.TweetRepository {
 }
 
 // Get returns tweets.
-func (r mysqlTweetRepository) Get(from time.Time, to time.Time) (res []domain.Tweet, err error) {
-	err = r.db.Where("created_at BETWEEN ? AND ?", from, to).Order("created_at desc").Find(&res).Error
+func (r mysqlTweetRepository) Get(user domain.User, from time.Time, to time.Time) ([]domain.Tweet, error) {
+	var tweetList []domain.Tweet
+	err := r.db.Where("user_id = ?", user.ID).Where("created_at BETWEEN ? AND ?", from, to).Order("created_at desc").Find(&tweetList).Error
+	if err != nil {
+		return nil, err
+	}
 
-	return
+	return tweetList, nil
 }
 
 // Create creates a new tweet.
-func (r mysqlTweetRepository) Create(comment string) (tweet domain.Tweet, err error) {
-	tweet = domain.Tweet{Comment: comment}
+func (r mysqlTweetRepository) Create(user domain.User, comment string) (tweet domain.Tweet, err error) {
+	tweet = domain.Tweet{
+		Comment: comment,
+		UserID:  user.ID,
+	}
 	err = r.db.Create(&tweet).Error
 
 	return
 }
 
-// Delete deletes a tweet specified by id.
-func (r mysqlTweetRepository) Delete(id int) error {
+// Delete deletes a tweet.
+func (r mysqlTweetRepository) Delete(user domain.User, id uint) error {
+	var tweet domain.Tweet
+	err := r.db.First(&tweet, id).Error
+	if err != nil {
+		return err
+	}
+
+	// HACK: Do in usecase
+	if tweet.UserID != user.ID {
+		return infrastructure.NoAuthorizationError{}
+	}
+
 	result := r.db.Delete(&domain.Tweet{}, id)
 	return result.Error
 }
